Feat: Review 삭제 권한 Exception 처리 추가

+++ src/main/java/kr/sujin/cafereview/service/ReviewDeleteService.java
@@ -1,12 +1,17 @@
 package kr.sujin.cafereview.service;
 
+import kr.sujin.cafereview.entity.Review;
 import kr.sujin.cafereview.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
 
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Optional;
+
 import javax.persistence.EntityNotFoundException;
+import javax.security.auth.message.AuthException;
 
 @Service
 @Transactional
@@ -14,9 +19,14 @@ import javax.persistence.EntityNotFoundException;
 public class ReviewDeleteService {
     private final ReviewRepository reviewRepository;
 
-    public void deleteReview(Long reviewId){
+    public void deleteReview(Long reviewId, String email){
         // 이미 존재하지 않는 리뷰 -> EntityNotFoundException
-        reviewRepository.getReviewById(reviewId).orElseThrow(EntityNotFoundException::new);;
+        Review review = reviewRepository.getReviewById(reviewId).orElseThrow(EntityNotFoundException::new);
+        if (review.getEmail() != email){
+            // 권한 없음
+            throw new AccessDeniedException("삭제 권한 없음");
+        }
+        
 
         // softDelete
         reviewRepository.deleteByReviewId(reviewId);
